migrate TestTallyNoQuorum to use simapp
