+++ x/gov/keeper/tally_test.go
@@ -3,6 +3,8 @@ package keeper_test
 import (
 	"testing"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	abci "github.com/tendermint/tendermint/abci/types"
 
 	"github.com/stretchr/testify/require"
@@ -31,4 +33,29 @@ func TestTallyNoOneVotes(t *testing.T) {
 	require.False(t, passes)
 	require.True(t, burnDeposits)
 	require.True(t, tallyResults.Equals(types.EmptyTallyResult()))
+}
+
+func TestTallyNoQuorum(t *testing.T) {
+	app := simapp.Setup(false)
+	ctx := app.BaseApp.NewContext(false, abci.Header{})
+
+	createValidators(ctx, app, []int64{2, 5, 0})
+
+	addrs := simapp.AddTestAddrsIncremental(app, ctx, 1, sdk.NewInt(10000000))
+
+	tp := TestProposal
+	proposal, err := app.GovKeeper.SubmitProposal(ctx, tp)
+	require.NoError(t, err)
+	proposalID := proposal.ProposalID
+	proposal.Status = types.StatusVotingPeriod
+	app.GovKeeper.SetProposal(ctx, proposal)
+
+	err = app.GovKeeper.AddVote(ctx, proposalID, addrs[0], types.OptionYes)
+	require.Nil(t, err)
+
+	proposal, ok := app.GovKeeper.GetProposal(ctx, proposalID)
+	require.True(t, ok)
+	passes, burnDeposits, _ := app.GovKeeper.Tally(ctx, proposal)
+	require.False(t, passes)
+	require.True(t, burnDeposits)
 }
